feat: add UserController and fix conflict with client service

+++ src/main/java/com/pablocos/KinalApp1/controller/ClienteController.java
@@ -101,7 +101,7 @@ public class ClienteController {
     }
 
     @GetMapping("/estado/{estado}")
-    public ResponseEntity<List<Cliente>> buscarPorEstado(@PathVariable int estado) {
+    public ResponseEntity<List<Cliente>> buscarPorEstado(@PathVariable Long estado) {
         List<Cliente> clientes = clienteService.buscarPorEstadoConFor(estado);
         if (clientes.isEmpty()) {
             return ResponseEntity.notFound().build();
